# chivospot/backend/openapi.yaml
openapi: 3.0.3
info:
  title: ChivoSpot API
  version: 1.0.0
  description: API para descubrimiento y gestión de eventos en El Salvador.
servers:
  - url: https://api.chivospot.sv
  - url: http://localhost:4000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        descripcion:
          type: string
        fecha_inicio:
          type: string
          format: date-time
        fecha_fin:
          type: string
          format: date-time
        precio:
          type: number
        estado_publicacion:
          type: string
...
paths:
...
  /events:
    get:
      summary: Buscar eventos
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: department_id
          schema:
            type: integer
        - in: query
          name: municipality_id
          schema:
            type: integer
        - in: query
          name: category_id
          schema:
            type: array
            items:
              type: integer
        - in: query
          name: bbox
          schema:
            type: string
        - in: query
          name: order
          schema:
            type: string
      responses:
        '200':
          description: Resultados
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
  /events/{id}:
    get:
      summary: Obtener evento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Evento
  /events/{id}/favorite:
    post:
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Marcado como favorito
    delete:
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Eliminado
...
security:
  - bearerAuth: []